SDG queries of the Aurora Universities Network

A dataset containing the SDG queries version 5.0 of the Aurora Universities Network. See the corresponding GitHub repository. For the actual implementation of the queries see aurora_simple, aurora_and, aurora_w, and the queries hard-coded in detect_aurora. There are multiple queries per SDG (one per row). In comparison to previous versions, this version of the queries Aurora added more keywords related to academic terminology to be able to detect more research papers related to the SDGs. The current version also drew inspiration from the SIRIS query system (siris_queries). The Aurora queries were designed to be precise rather than sensitive. To achieve this the queries make use complex keyword-combinations using several different logical search operators. All SDGs (1-17) are covered.

aurora_queries

Format

A data frame with 378 rows and 5 columns

system

Name of system

sdg

Label of the SDG

sdg_title

Title of the SDG

sdg_description

Description of the SDG

query_id

Index of the query

query

Original SDG query

Source

https://github.com/Aurora-Network-Global/sdg-queries/releases/tag/v5.0