Detecting UN Sustainable Development Goals in Text

The text2sdg package provides functions for detecting SDGs in text, as well as for analyzing and visualization the hits found in text. The following provides a brief overview of the contents of the package.

Detect functions

detect_sdg detects SDGs in text using up to five different query systems: Aurora, Elsevier, SIRIS, SDSN, and Ontology.

detect_any detects SDGs in text using self-specified queries utilizing the lucene-style syntax of the corpustools package.

Analysis functions

plot_sdg visualizes the relative frequency of SDG hits across query systems.

crosstab_sdg calculates cross tables of correlations between either the query systems or the different SDGs.

Datasets

projects contain random selection of research project descriptions from the P3 database of the Swiss National Science Foundation.

aurora_queries, elsevier_queries, siris_queries, sdsn_queries, and ontology_queries contain a mapping of SDGs and search queries as they are employed in the respective systems.

Examples

if (FALSE) {
# detect SDGs using default systems
hits <- detect_sdg(projects)

#' # detect SDGs using all five systems
hits <- detect_sdg(projects,
                   system = c("aurora", "elsevier", "siris", "sdsn", "ontology"))

# visualize SDG frequencies
plot_sdg(hits)

# correlations between systems
crosstab_sdg(hits)

# correlations between SDGs
crosstab_sdg(hits, compare = "sdgs")
}