Compare query systems and SDGs

crosstab_sdg calculates cross tables (aka contingency tables) of SGSs or systems across hits identified via detect_sdg.

crosstab_sdg(hits, compare = c("systems", "sdgs"), systems = NULL, sdgs = NULL)

Arguments

hits
data frame as returned by detect_sdg. Must include columns document, sdg, system, and hit.

compare
character specifying whether systems or SDGs should be cross tabulated.

systems
character vector specifying the query systems to be cross tabulated. Values must be available in the system column of hits. systems of length greater 1 result, by default, in a stacked barplot. Defaults to NULL in which case available values are retrieved from hits.

sdgs
numeric vector with integers between 1 and 17 specifying the SDGs to be cross tabluated. Values must be available in the sdg column of hits. Defaults to NULL in which case available values are retrieved from hits.

Value

matrix showing correlation coefficients for all pairs of query systems (if compare = "systems") or SDGs (if compare = "SDGs").

Details

crosstab_sdg determines correlations between either query systems or SDGs. The respectively other dimension will be ignored. Note that correlations between SDGs may vary between query systems.

Examples

# \donttest{
# run sdg detection
hits <- detect_sdg(projects)

#> 
#> Running aurora queries
#> Running siris queries
#> Running elsevier queries

# create cross table of systems
crosstab_sdg(hits)

#>             Aurora  Elsevier     SIRIS
#> Aurora   1.0000000 0.2434814 0.3136179
#> Elsevier 0.2434814 1.0000000 0.2690456
#> SIRIS    0.3136179 0.2690456 1.0000000

# create cross table of systems
crosstab_sdg(hits, compare = "sdgs")

#>              SDG-01       SDG-02       SDG-03       SDG-04       SDG-05
#> SDG-01  1.000000000 -0.005193421  0.074185067 -0.005193421 -0.004020119
#> SDG-02 -0.005193421  1.000000000  0.011068212 -0.003344482 -0.002588893
#> SDG-03  0.074185067  0.011068212  1.000000000 -0.022814069 -0.017659893
#> SDG-04 -0.005193421 -0.003344482 -0.022814069  1.000000000 -0.002588893
#> SDG-05 -0.004020119 -0.002588893 -0.017659893 -0.002588893  1.000000000
#> SDG-06  0.140468432 -0.002990395 -0.020398702 -0.002990395 -0.002314802
#> SDG-07 -0.008716533  0.234957114  0.002328853 -0.005613310 -0.004345146
#> SDG-08  0.084604578 -0.004737741 -0.032318054 -0.004737741  0.179701902
#> SDG-09 -0.005193421 -0.003344482 -0.022814069 -0.003344482 -0.002588893
#> SDG-10  0.061082651 -0.006191820 -0.042236918  0.103075584  0.136176391
#> SDG-11  0.280591424 -0.007529019 -0.020848500 -0.007529019 -0.005828056
#> SDG-12 -0.005691012 -0.003664922 -0.024999923 -0.003664922 -0.002836939
#> SDG-13 -0.015241757  0.060294917 -0.019603569 -0.009815452 -0.007597936
#> SDG-14  0.079992693  0.130592597 -0.033906841 -0.004970652 -0.003847678
#> SDG-15 -0.011691309  0.173167440 -0.020848500 -0.007529019 -0.005828056
#> SDG-16  0.041740251 -0.008120019 -0.041206797 -0.008120019 -0.006285537
#> SDG-17 -0.004020119 -0.002588893  0.026052713 -0.002588893 -0.002004008
#>              SDG-06       SDG-07       SDG-08       SDG-09       SDG-10
#> SDG-01  0.140468432 -0.008716533  0.084604578 -0.005193421  0.061082651
#> SDG-02 -0.002990395  0.234957114 -0.004737741 -0.003344482 -0.006191820
#> SDG-03 -0.020398702  0.002328853 -0.032318054 -0.022814069 -0.042236918
#> SDG-04 -0.002990395 -0.005613310 -0.004737741 -0.003344482  0.103075584
#> SDG-05 -0.002314802 -0.004345146  0.179701902 -0.002588893  0.136176391
#> SDG-06  1.000000000 -0.005019019  0.154619381 -0.002990395 -0.005536280
#> SDG-07 -0.005019019  1.000000000  0.077245351 -0.005613310 -0.010392224
#> SDG-08  0.154619381  0.077245351  1.000000000 -0.004737741 -0.008771235
#> SDG-09 -0.002990395 -0.005613310 -0.004737741  1.000000000 -0.006191820
#> SDG-10 -0.005536280 -0.010392224 -0.008771235 -0.006191820  1.000000000
#> SDG-11  0.094246717 -0.012636553  0.181313350 -0.007529019 -0.013938880
#> SDG-12 -0.003276910  0.103690485  0.124600112 -0.003664922 -0.006785068
#> SDG-13  0.069583301  0.067577244  0.135071469 -0.009815452 -0.018171876
#> SDG-14 -0.004444401 -0.008342642 -0.007041349  0.130592597 -0.009202438
#> SDG-15  0.094246717 -0.012636553  0.053327456 -0.007529019 -0.013938880
#> SDG-16 -0.007260338 -0.013628475  0.047993995 -0.008120019  0.213664193
#> SDG-17 -0.002314802 -0.004345146 -0.003667386 -0.002588893 -0.004792958
#>              SDG-11       SDG-12       SDG-13       SDG-14       SDG-15
#> SDG-01  0.280591424 -0.005691012 -0.015241757  0.079992693 -0.011691309
#> SDG-02 -0.007529019 -0.003664922  0.060294917  0.130592597  0.173167440
#> SDG-03 -0.020848500 -0.024999923 -0.019603569 -0.033906841 -0.020848500
#> SDG-04 -0.007529019 -0.003664922 -0.009815452 -0.004970652 -0.007529019
#> SDG-05 -0.005828056 -0.002836939 -0.007597936 -0.003847678 -0.005828056
#> SDG-06  0.094246717 -0.003276910  0.069583301 -0.004444401  0.094246717
#> SDG-07 -0.012636553  0.103690485  0.067577244 -0.008342642 -0.012636553
#> SDG-08  0.181313350  0.124600112  0.135071469 -0.007041349  0.053327456
#> SDG-09 -0.007529019 -0.003664922 -0.009815452  0.130592597 -0.007529019
#> SDG-10 -0.013938880 -0.006785068 -0.018171876 -0.009202438 -0.013938880
#> SDG-11  1.000000000 -0.008250387  0.104168348  0.049845549  0.064406780
#> SDG-12 -0.008250387  1.000000000 -0.010755886 -0.005446899 -0.008250387
#> SDG-13  0.104168348 -0.010755886  1.000000000  0.127502651  0.261999179
#> SDG-14  0.049845549 -0.005446899  0.127502651  1.000000000  0.110880915
#> SDG-15  0.064406780 -0.008250387  0.261999179  0.110880915  1.000000000
#> SDG-16  0.095179979 -0.008898012  0.005517475 -0.012068177 -0.018279599
#> SDG-17 -0.005828056 -0.002836939 -0.007597936 -0.003847678 -0.005828056
#>              SDG-16       SDG-17
#> SDG-01  0.041740251 -0.004020119
#> SDG-02 -0.008120019 -0.002588893
#> SDG-03 -0.041206797  0.026052713
#> SDG-04 -0.008120019 -0.002588893
#> SDG-05 -0.006285537 -0.002004008
#> SDG-06 -0.007260338 -0.002314802
#> SDG-07 -0.013628475 -0.004345146
#> SDG-08  0.047993995 -0.003667386
#> SDG-09 -0.008120019 -0.002588893
#> SDG-10  0.213664193 -0.004792958
#> SDG-11  0.095179979 -0.005828056
#> SDG-12 -0.008898012 -0.002836939
#> SDG-13  0.005517475 -0.007597936
#> SDG-14 -0.012068177 -0.003847678
#> SDG-15 -0.018279599 -0.005828056
#> SDG-16  1.000000000 -0.006285537
#> SDG-17 -0.006285537  1.000000000

# }